refactored project, deleted old home_screen.dart file and replaced it with the new home_screen.dart file. moved file into screens folder
